Refactor LLM-DSPy integration and enhance module registration
- Consolidated the LLM-DSPy framework by restructuring the `__init__.py` file to improve organization and clarity.
- Introduced new utility functions and classes for better modularity, including `LLMRetriever`, `QueryTransformer`, and `ContextRewriter`.
- Updated the command registration process to dynamically handle input fields and improve error handling for module signatures.
- Enhanced the overall architecture to support Retrieval-Augmented Generation (RAG) capabilities more effectively.
- Improved documentation and versioning for better clarity on the framework's functionality.
