+++ llm_dspy/cli/__init__.py
@@ -1,0 +1,5 @@
+"""CLI package for LLM-DSPy."""
+
+from .commands import register_commands
+
+__all__ = ['register_commands']
