+++ llm_dspy/core/__init__.py
@@ -1,0 +1,5 @@
+"""Core functionality for LLM-DSPy."""
+
+from .module import run_dspy_module
+
+__all__ = ['run_dspy_module']
