+++ llm_dspy/rag/__init__.py
@@ -1,0 +1,7 @@
+"""RAG (Retrieval-Augmented Generation) functionality for LLM-DSPy."""
+
+from .retriever import LLMRetriever
+from .transformer import QueryTransformer
+from .enhanced import EnhancedRAGModule, ContextRewriter
+
+__all__ = ['LLMRetriever', 'QueryTransformer', 'EnhancedRAGModule', 'ContextRewriter']
